+++ app/controllers/api/trainees/placements_controller.rb
@@ -45,7 +45,7 @@ module Api
       def slug = params[:slug]
 
       def trainee_serializer_klass
-        Serializer.for(model: :trainee, version: version)
+        Api::Serializer.for(model: :trainee, version: version)
       end
 
       def model = :placement
