+++ spec/requests/api/v0.1/trainees/put_degrees_spec.rb
@@ -119,7 +119,7 @@ describe "`PUT /trainees/:trainee_slug/degrees/:slug` endpoint" do
       let(:degrees_attributes) do
         non_uk_degree.attributes.symbolize_keys.slice(
           :country, :grade, :grade_uuid, :subject, :subject_uuid, :institution, :institution_uuid, :uk_degree, :uk_degree_uuid, :graduation_year, :non_uk_degree
-        ).merge(DegreeSerializer::V01.new(non_uk_degree).as_hash.symbolize_keys)
+        ).merge(Api::DegreeSerializer::V01.new(non_uk_degree).as_hash.symbolize_keys)
       end
 
       before do
