+++ spec/serializers/api/placement_serializer/v01_spec.rb
@@ -2,7 +2,7 @@
 
 require "rails_helper"
 
-RSpec.describe PlacementSerializer::V01 do
+RSpec.describe Api::PlacementSerializer::V01 do
   shared_examples_for "a placement serialiser" do
     %w[
       placement_id
