+++ spec/serializers/api/serializer_spec.rb
@@ -2,12 +2,12 @@
 
 require "rails_helper"
 
-RSpec.describe Serializer do
-  serializers = %i[degree hesa_trainee_detail placement trainee]
+RSpec.describe Api::Serializer do
+  serializers = %w[degree hesa_trainee_detail placement trainee]
   context "v0.1" do
     serializers.each do |serializer|
       it "#{serializer} has been implemented" do
-        expect(described_class.for(model: serializer, version: "v0.1")).to be(Object.const_get("#{serializer}_Serializer::V01".camelize))
+        expect(described_class.for(model: serializer, version: "v0.1")).to be(Object.const_get("Api::#{serializer.camelize}Serializer::V01"))
       end
     end
   end
@@ -15,7 +15,7 @@ RSpec.describe Serializer do
   context "v1.0" do
     serializers.each do |serializer|
       it "#{serializer} has not been implemented" do
-        expect { described_class.for(model: serializer, version: "v1.0") }.to raise_error(NotImplementedError, "#{serializer}_Serializer::V10".camelize)
+        expect { described_class.for(model: serializer, version: "v1.0") }.to raise_error(NotImplementedError, "Api::#{serializer.camelize}Serializer::V10")
       end
     end
   end
@@ -27,7 +27,7 @@ RSpec.describe Serializer do
 
     serializers.each do |serializer|
       it "#{serializer} has not been implemented" do
-        expect { described_class.for(model: serializer, version: "v0.1") }.to raise_error(NotImplementedError, "#{serializer}_Serializer::V01".camelize)
+        expect { described_class.for(model: serializer, version: "v0.1") }.to raise_error(NotImplementedError, "Api::#{serializer.camelize}Serializer::V01")
       end
     end
   end
