+++ spec/services/api/trainees/withdraw_response_spec.rb
@@ -34,7 +34,7 @@ describe Api::Trainees::WithdrawResponse do
     end
 
     it "uses the trainee serializer" do
-      expect(TraineeSerializer::V01).to receive(:new).with(trainee).and_return(double(as_hash: trainee.attributes)).at_least(:once)
+      expect(Api::TraineeSerializer::V01).to receive(:new).with(trainee).and_return(double(as_hash: trainee.attributes)).at_least(:once)
 
       subject
     end
